msg/async: transmit V2 messages with new preable format.

+++ src/msg/async/ProtocolV2.h
@@ -47,7 +47,7 @@ private:
   }
 
 public:
-  enum class Tag : uint32_t {
+  enum class Tag : __u8 {
     HELLO = 1,
     AUTH_REQUEST,
     AUTH_BAD_METHOD,
